crypto: ccp: move vpsp-related functions to vpsp.c
hygon inclusion
category: feature
CVE: NA

---------------------------

Previously, vpsp-related functions or data structure
definitions were scattered across different files.

To improve manageability and reduce conflicts
during patch porting, vpsp-related content is now
consolidated into psp-dev.c as much as possible.
